pyproject.toml
[project]
name = "weather-stream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
